remove compiler warnings
git-svn-id: https://openhr20.svn.sourceforge.net/svnroot/openhr20@278 e6282f77-c345-0410-9e5f-9d7f1dee61d8

+++ rfmsrc/OpenHR20/main.c
@@ -334,9 +334,9 @@ int __attribute__ ((noreturn)) main(void)
 
 FUSES = 
 {
-    .low = (CKSEL0 & CKSEL2 & CKSEL3 & SUT0 & CKDIV8),  //0x62
-    .high = (BOOTSZ0 & BOOTSZ1 & EESAVE & SPIEN & JTAGEN),
-    .extended = (BODLEVEL0),
+    .low = (uint8_t)(CKSEL0 & CKSEL2 & CKSEL3 & SUT0 & CKDIV8),  //0x62
+    .high = (uint8_t)(BOOTSZ0 & BOOTSZ1 & EESAVE & SPIEN & JTAGEN),
+    .extended = (uint8_t)(BODLEVEL0),
 };
 
 /*!
